combine hash into nmt
